pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic-settings = "^2.7.0"
aiofiles = "^24.1.0"
asyncpg = "^0.30.0"
minio = "^7.2.13"
passlib = {extras = ["argon2"], version = "^1.7.4"}
pyjwt = "^2.10.1"
bs4 = "^0.0.2"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
